Extract useInactivityAutoSubmit from App.tsx
Move the 60s no-activity auto-submit timer for open sub-menus into a
hook. The hook captures the submit and reset callbacks via refs so
their identity changing on every render does not re-arm the timer —
only navStack changes do, matching the prior `[navStack]` dep list.

53/53 tests pass.
